Move preference converters into their own package
