+++ app/src/main/java/org/simple/clinic/patient/sync/PatientSyncModule.kt
@@ -6,8 +6,8 @@ import dagger.Module
 import dagger.Provides
 import org.simple.clinic.util.None
 import org.simple.clinic.util.Optional
-import org.simple.clinic.util.OptionalRxPreferencesConverter
-import org.simple.clinic.util.StringPreferenceConverter
+import org.simple.clinic.util.preference.OptionalRxPreferencesConverter
+import org.simple.clinic.util.preference.StringPreferenceConverter
 import retrofit2.Retrofit
 import javax.inject.Named
