+++ app/src/main/java/org/simple/clinic/util/preference/LocalePreferenceConverter.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.util
+package org.simple.clinic.util.preference
 
 import com.f2prateek.rx.preferences2.Preference
 import java.util.Locale
