+++ app/src/main/java/org/simple/clinic/util/preference/OptionalRxPreferencesConverter.kt
@@ -1,6 +1,9 @@
-package org.simple.clinic.util
+package org.simple.clinic.util.preference
 
 import com.f2prateek.rx.preferences2.Preference
+import org.simple.clinic.util.Just
+import org.simple.clinic.util.None
+import org.simple.clinic.util.Optional
 
 class OptionalRxPreferencesConverter<T : Any>(private val valueConverter: Preference.Converter<T>) : Preference.Converter<Optional<T>> {
